Sat, Feb 17, 2024, 11:07 PM +05:30

+++ readability/readability.c
@@ -27,7 +27,7 @@ int main()
     }
     else if (index >= 1 && index < 16)
     {
-        printf("grade %i",round(index));
+        printf("grade %i",int(round(index)));
     }
     else if (index > 16)
     {
